Factor out station-info functions
